+++ src/main/java/org/prgrms/kdt/model/customer/Customer.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt.model;
+package org.prgrms.kdt.model.customer;
 
 import java.util.UUID;
