+++ src/main/java/org/prgrms/kdt/model/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt.model;
+package org.prgrms.kdt.model.order;
 
 public enum OrderStatus {
     ACCEPTED,
