+++ src/main/java/org/prgrms/kdt/model/wallet/Wallet.java
@@ -1,0 +1,5 @@
+package org.prgrms.kdt.model.wallet;
+
+public class Wallet {
+
+}
