+++ src/main/java/org/prgrms/kdt/repository/JdbcCustomerRepository.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import org.prgrms.kdt.model.Customer;
-import org.prgrms.kdt.model.CustomerType;
+import org.prgrms.kdt.model.customer.Customer;
+import org.prgrms.kdt.model.customer.CustomerType;
 import org.prgrms.kdt.util.UuidUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
