--- src/main/java/org/prgrms/kdt/repository/OrderRepository.java
@@ -1,8 +1,0 @@
-package org.prgrms.kdt.repository;
-
-import org.prgrms.kdt.model.Order;
-
-public interface OrderRepository {
-
-    Order insert(Order order);
-}
