+++ src/main/java/org/prgrms/kdt/repository/customer/FileRepository.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt.repository;
+package org.prgrms.kdt.repository.customer;
 
 import java.io.IOException;
