+++ src/main/java/org/prgrms/kdt/repository/order/OrderRepository.java
@@ -1,0 +1,8 @@
+package org.prgrms.kdt.repository.order;
+
+import org.prgrms.kdt.model.order.Order;
+
+public interface OrderRepository {
+
+    Order insert(Order order);
+}
