+++ src/main/java/org/prgrms/kdt/service/OrderService.java
@@ -1,8 +1,8 @@
 package org.prgrms.kdt.service;
 
-import org.prgrms.kdt.model.Order;
-import org.prgrms.kdt.model.OrderItem;
-import org.prgrms.kdt.repository.OrderRepository;
+import org.prgrms.kdt.model.order.Order;
+import org.prgrms.kdt.model.order.OrderItem;
+import org.prgrms.kdt.repository.order.OrderRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
